RavenDB-24796 Cleanup
